Moved celestial focus to camera plugin
